requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.1
django-cors-headers==4.1.0
django-environ==0.10.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
whitenoise==6.5.0
